add a column name mapper func
it may be generally helpful to process column names
when mapping, similar to how field names and tag names can be preprocessed in the mapper.

E.g. matching go PascalCase field names to sql camelCased column names. This could be achived with custom db tags but it feels much dryer to pass in one mapper func than have to redundantly add a tag to every struct field to change a single character.